backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0

# LangChain packages - v1.0 compatible
langchain==1.0.0
langgraph==1.0.1
langchain-aws==1.0.0
langchain-core>=1.0.0
langchain-community>=0.4

# LangChain provider integrations
langchain-mistralai>=0.1.0

# AI provider SDKs (using direct SDKs instead of langchain wrappers)
google-generativeai>=0.8.0
boto3>=1.34.72
httpx>=0.27.0
